+++ packages/tiny-react/src/index.ts
@@ -1,6 +1,6 @@
 export { render, hydrate } from "./reconciler";
-export type { VNode, FC } from "./virtual-dom";
-export { Fragment, h, createElement } from "./helper/hyperscript";
+export { type VNode, type FC, Fragment, createElement } from "./virtual-dom";
+export { h } from "./helper/hyperscript";
 export type { ComponentChildren, ComponentChild } from "./helper/common";
 export * from "./hooks";
 export * from "./compat";
